requirements.txt
aiogram==3.0.0rc2
pydantic==2.2.1
python-dotenv==1.0.0
ffmpeg-python==0.2.0